backend/requirements.txt
langchain
langchain-google-genai
google-generativeai
dotenv
openai
langchain-community
wave
google-genai
pillow
langgraph